fw-6473, move base tests into own package
