+++ firstvoices/backend/tests/test_apis/test_category_api.py
@@ -17,7 +17,9 @@ from backend.tests.factories import (
     get_app_admin,
     get_non_member_user,
 )
-from backend.tests.test_apis.base_api_test import BaseUncontrolledSiteContentApiTest
+from backend.tests.test_apis.base.base_api_test import (
+    BaseUncontrolledSiteContentApiTest,
+)
 from backend.tests.utils import find_object_by_id
