+++ firstvoices/backend/tests/test_apis/test_documents_api.py
@@ -5,7 +5,7 @@ import pytest
 from backend.models.constants import Visibility
 from backend.models.media import Document
 from backend.tests import factories
-from backend.tests.test_apis.base_media_test import BaseMediaApiTest
+from backend.tests.test_apis.base.base_media_test import BaseMediaApiTest
 
 
 class TestDocumentEndpoint(BaseMediaApiTest):
