+++ firstvoices/backend/tests/test_apis/test_immersion_label_api.py
@@ -5,7 +5,9 @@ import pytest
 from backend.models.constants import AppRole, Role, Visibility
 from backend.models.immersion_labels import ImmersionLabel
 from backend.tests import factories
-from backend.tests.test_apis.base_api_test import BaseUncontrolledSiteContentApiTest
+from backend.tests.test_apis.base.base_api_test import (
+    BaseUncontrolledSiteContentApiTest,
+)
 from backend.tests.test_apis.test_dictionary_api import (
     assert_dictionary_entry_summary_response,
 )
