+++ firstvoices/backend/tests/test_apis/test_search_apis/test_languages_api.py
@@ -10,8 +10,8 @@ from backend.tests.factories.access import (
     get_non_member_user,
     get_superadmin,
 )
-from backend.tests.test_apis.base_api_test import BaseApiTest
-from backend.tests.test_apis.base_media_test import MediaTestMixin
+from backend.tests.test_apis.base.base_api_test import BaseApiTest
+from backend.tests.test_apis.base.base_media_test import MediaTestMixin
 from backend.tests.test_apis.test_search_apis.base_search_test import SearchMocksMixin
