+++ firstvoices/backend/tests/test_apis/test_storypage_api.py
@@ -6,9 +6,8 @@ from rest_framework.reverse import reverse
 from backend.models.constants import Role, Visibility
 from backend.models.story import StoryPage
 from backend.tests import factories
-
-from .base_api_test import BaseControlledSiteContentApiTest
-from .base_media_test import RelatedMediaTestMixin
+from backend.tests.test_apis.base.base_api_test import BaseControlledSiteContentApiTest
+from backend.tests.test_apis.base.base_media_test import RelatedMediaTestMixin
 
 
 class TestStoryPageEndpoint(RelatedMediaTestMixin, BaseControlledSiteContentApiTest):
