+++ firstvoices/backend/tests/test_models/test_media_models.py
@@ -25,7 +25,10 @@ from backend.models.media import (
 )
 from backend.tests import factories
 from backend.tests.factories.media_factories import get_video_content
-from backend.tests.test_apis.base_media_test import VIMEO_VIDEO_LINK, YOUTUBE_VIDEO_LINK
+from backend.tests.test_apis.base.base_media_test import (
+    VIMEO_VIDEO_LINK,
+    YOUTUBE_VIDEO_LINK,
+)
 
 logger = logging.getLogger(__name__)
